+++ classes/graphics/modifiers/fontsize.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -39,12 +39,12 @@ S::GUI::Rect S::GUI::FontSizeModifier::TranslateRect(const Rect &r) const
 	return r * fontSize / 96;
 }
 
-S::Void S::GUI::FontSizeModifier::SetFontSize(Int nFontSize)
+S::Void S::GUI::FontSizeModifier::SetFontSize(Short nFontSize)
 {
 	fontSize = nFontSize;
 }
 
-S::Int S::GUI::FontSizeModifier::GetFontSize() const
+S::Short S::GUI::FontSizeModifier::GetFontSize() const
 {
 	return fontSize;
 }
