+++ classes/gui/mdi/window.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -10,7 +10,7 @@
 
 #include <smooth/gui/mdi/window.h>
 
-const S::Int	 S::GUI::MDI::Window::classID = S::Object::RequestClassID();
+const S::Short	 S::GUI::MDI::Window::classID = S::Object::RequestClassID();
 
 S::GUI::MDI::Window::Window(const String &title, const Point &iPos, const Size &iSize) : GUI::Window(title, iPos, iSize)
 {
@@ -21,7 +21,7 @@ S::GUI::MDI::Window::~Window()
 {
 }
 
-S::Bool S::GUI::MDI::Window::IsTypeCompatible(Int compType) const
+S::Bool S::GUI::MDI::Window::IsTypeCompatible(Short compType) const
 {
 	if (compType == Object::classID || compType == Widget::classID || compType == Window::classID)	return True;
 	else												return False;
