+++ classes/gui/widgets/basic/titlebar.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -16,7 +16,7 @@
 #include <smooth/gui/application/application.h>
 #include <smooth/misc/binary.h>
 
-const S::Int	 S::GUI::Titlebar::classID = S::Object::RequestClassID();
+const S::Short	 S::GUI::Titlebar::classID = S::Object::RequestClassID();
 
 S::GUI::Titlebar::Titlebar(Int buttons) : Widget(Point(), Size(0, 19))
 {
