+++ classes/gui/widgets/hotspot/simplebutton.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -11,7 +11,7 @@
 #include <smooth/gui/widgets/hotspot/simplebutton.h>
 #include <smooth/graphics/surface.h>
 
-const S::Int	 S::GUI::HotspotSimpleButton::classID = S::Object::RequestClassID();
+const S::Short	 S::GUI::HotspotSimpleButton::classID = S::Object::RequestClassID();
 
 S::GUI::HotspotSimpleButton::HotspotSimpleButton(const Point &iPos, const Size &iSize) : Hotspot(iPos, iSize)
 {
