+++ classes/gui/widgets/layer.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -14,7 +14,7 @@
 #include <smooth/gui/widgets/basic/tabwidget.h>
 #include <smooth/graphics/surface.h>
 
-const S::Int	 S::GUI::Layer::classID = S::Object::RequestClassID();
+const S::Short	 S::GUI::Layer::classID = S::Object::RequestClassID();
 
 S::GUI::Layer::Layer(const String &name) : Widget(Point(0, 0), Size(32768, 32768))
 {
