+++ classes/gui/widgets/multi/image/imageentry.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -12,7 +12,7 @@
 #include <smooth/gui/widgets/basic/image.h>
 #include <smooth/graphics/surface.h>
 
-const S::Int	 S::GUI::ImageEntry::classID = S::Object::RequestClassID();
+const S::Short	 S::GUI::ImageEntry::classID = S::Object::RequestClassID();
 
 S::GUI::ImageEntry::ImageEntry(const Bitmap &iBitmap, const Size &iSize) : ListEntry(NIL)
 {
@@ -86,7 +86,7 @@ S::Void S::GUI::ImageEntry::OnChangeSize(const Size &newSize)
 	image->SetSize(newSize - Size(4, 4));
 }
 
-S::Bool S::GUI::ImageEntry::IsTypeCompatible(Int compType) const
+S::Bool S::GUI::ImageEntry::IsTypeCompatible(Short compType) const
 {
 	if (compType == Object::classID || compType == Widget::classID || compType == ListEntry::classID)	return True;
 	else													return False;
