+++ classes/gui/widgets/multi/list/listentryseparator.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -12,7 +12,7 @@
 #include <smooth/gui/widgets/hotspot/hotspot.h>
 #include <smooth/graphics/surface.h>
 
-const S::Int	 S::GUI::ListEntrySeparator::classID = S::Object::RequestClassID();
+const S::Short	 S::GUI::ListEntrySeparator::classID = S::Object::RequestClassID();
 
 S::GUI::ListEntrySeparator::ListEntrySeparator() : ListEntry(NIL)
 {
@@ -53,7 +53,7 @@ S::Int S::GUI::ListEntrySeparator::Paint(Int message)
 	return Widget::Paint(message);
 }
 
-S::Bool S::GUI::ListEntrySeparator::IsTypeCompatible(Int compType) const
+S::Bool S::GUI::ListEntrySeparator::IsTypeCompatible(Short compType) const
 {
 	if (compType == Object::classID	   ||
 	    compType == Widget::classID	   ||
