+++ classes/gui/widgets/multi/menu/menu.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -10,7 +10,7 @@
 
 #include <smooth/gui/widgets/multi/menu/menu.h>
 
-const S::Int	 S::GUI::Menu::classID = S::Object::RequestClassID();
+const S::Short	 S::GUI::Menu::classID = S::Object::RequestClassID();
 
 S::GUI::Menu::Menu() : Widget(Point(), Size())
 {
