+++ classes/gui/widgets/multi/menu/popupmenu.cpp
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -19,7 +19,7 @@
 #include <smooth/graphics/surface.h>
 #include <smooth/misc/math.h>
 
-const S::Int	 S::GUI::PopupMenu::classID = S::Object::RequestClassID();
+const S::Short	 S::GUI::PopupMenu::classID = S::Object::RequestClassID();
 
 S::Signal1<S::Void, S::Int>	 S::GUI::PopupMenu::internalOnOpenPopupMenu;
